+++ front/app/api/navbar/useAddNavbarItem.test.ts
@@ -1,45 +1,43 @@
 import { renderHook, act } from '@testing-library/react-hooks';
 
-import useAddCause from './useAddCause';
+import useAddNavbarItem from './useAddNavbarItem';
 
 import { setupServer } from 'msw/node';
 import { rest } from 'msw';
 
 import createQueryClientWrapper from 'utils/testUtils/queryClientWrapper';
-import { causesData } from './__mocks__/useCauses';
+import { navbarItemsData } from './__mocks__/useNavbarItems';
 
 const apiPath = '*causes';
 
 const server = setupServer(
   rest.post(apiPath, (_req, res, ctx) => {
-    return res(ctx.status(200), ctx.json({ data: causesData[0] }));
+    return res(ctx.status(200), ctx.json({ data: navbarItemsData[0] }));
   })
 );
 
-describe('useAddCause', () => {
+describe('useAddNavbarItem', () => {
   beforeAll(() => server.listen());
   afterAll(() => server.close());
 
   it('mutates data correctly', async () => {
-    const { result, waitFor } = renderHook(() => useAddCause(), {
+    const { result, waitFor } = renderHook(() => useAddNavbarItem(), {
       wrapper: createQueryClientWrapper(),
     });
 
     act(() => {
       result.current.mutate({
-        title_multiloc: {
+        type: 'default_item',
+        navbarTitleMultiloc: {
           en: 'test',
         },
-        description_multiloc: {
-          en: 'test',
-        },
-        participation_context_id: '1',
-        participation_context_type: 'Project',
+        navbarCode: 'home',
+        navbarId: '1',
       });
     });
 
     await waitFor(() => expect(result.current.isSuccess).toBe(true));
-    expect(result.current.data?.data).toEqual(causesData[0]);
+    expect(result.current.data?.data).toEqual(navbarItemsData[0]);
   });
 
   it('returns error correctly', async () => {
@@ -49,22 +47,21 @@ describe('useAddCause', () => {
       })
     );
 
-    const { result, waitFor } = renderHook(() => useAddCause(), {
+    const { result, waitFor } = renderHook(() => useAddNavbarItem(), {
       wrapper: createQueryClientWrapper(),
     });
 
     act(() => {
       result.current.mutate({
-        title_multiloc: {
+        type: 'default_item',
+        navbarTitleMultiloc: {
           en: 'test',
         },
-        description_multiloc: {
-          en: 'test',
-        },
-        participation_context_id: '1',
-        participation_context_type: 'Project',
+        navbarCode: 'home',
+        navbarId: '1',
       });
     });
+
     await waitFor(() => expect(result.current.isError).toBe(true));
     expect(result.current.error).toBeDefined();
   });
