--- front/app/hooks/__mocks__/useNavbarItems.ts
@@ -1,3 +1,0 @@
-import navbarItems from 'services/__mocks__/navbarItems';
-
-export default jest.fn(() => navbarItems);
